# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: E-Commerce API V1
  version: 1.0.0
  description: |
    Mobil ve admin için REST API. Tüm müşteri endpoint'leri `/api/v1` altında.
    Admin `/api/v1/admin/*`. Locale `?locale=tr` ile.
servers:
  - url: https://api.example.com/api/v1
    description: Production
...
paths:
...
  /me:
    get:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: Profil
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: Profil güncelle
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
                locale: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /me/preferences:
    patch:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: Bildirim tercihi vb.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pushEnabled: { type: boolean }
      responses:
        '200':
          description: OK
  /me/device:
    post:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: FCM token kaydı
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fcmToken, platform]
              properties:
                fcmToken: { type: string }
                platform: { type: string, enum: [ios, android] }
      responses:
        '204':
          description: OK
...
  /products:
    get:
      tags: [Products]
      summary: Ürün listesi (cursor pagination)
      parameters:
        - $ref: '#/components/parameters/Locale'
        - name: categoryId
          in: query
          schema: { type: string }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{id}:
    get:
      tags: [Products]
      summary: Ürün detay
      parameters:
        - $ref: '#/components/parameters/Locale'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Not found
...
  /orders:
    post:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      summary: Sipariş oluştur + ödeme oturumu (idempotent)
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
                  payment:
                    type: object
                    properties:
                      redirectUrl: { type: string }
                      provider: { type: string }
        '409':
          description: Stok / fiyat değişti

    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  nextCursor: { type: string, nullable: true }
  /orders/{id}:
    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '404':
          description: Not found
...
components:
  parameters:
    Locale:
      name: locale
      in: query
      schema:
        type: string
        default: tr
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        phone: { type: string, nullable: true }
        locale: { type: string }
...
    ProductSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        thumbnailUrl: { type: string, nullable: true }
        minPrice: { type: number }
        currency: { type: string }
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
        nextCursor: { type: string, nullable: true }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          properties:
            description: { type: string }
            images:
              type: array
              items:
                type: object
                properties:
                  url: { type: string }
                  alt: { type: string, nullable: true }
            variants:
              type: array
              items:
                $ref: '#/components/schemas/Variant'
    Variant:
      type: object
      properties:
        id: { type: string }
        sku: { type: string }
        price: { type: number }
        stock: { type: integer }
        attributes: { type: object, additionalProperties: true }
...
    CreateOrderRequest:
      type: object
      required: [addressId, couponCode]
      properties:
        addressId: { type: string }
        couponCode: { type: string, nullable: true }
        clientRequestId: { type: string, description: Idempotency body fallback }
    Order:
      type: object
      properties:
        id: { type: string }
        status:
          type: string
          enum:
            [
              pending_payment,
              paid,
              processing,
              shipped,
              delivered,
              cancelled,
            ]
        paymentStatus:
          type: string
        total: { type: number }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                properties:
                  variantId: { type: string }
                  productName: { type: string }
                  quantity: { type: integer }
                  unitPrice: { type: number }
...